Fix hash table warnings and create CN error message struct
Stop C hash table implementation from throwing warnings by making underlying integer type (signed long) uniform throughout. Create struct for carrying around info for reporting CN runtime errors - currently unused.

+++ executable-spec/hash_table.h
@@ -29,7 +29,7 @@ SOFTWARE.
 
 // Hash table entry (slot may be filled or empty).
 typedef struct {
-    unsigned int *key;  // key is NULL if this slot is empty
+    signed long *key;  // key is NULL if this slot is empty
     void* value;
 } ht_entry;
